requirements.txt
Flask>=2.0
requests>=2.28
google-auth>=2.0
google-auth-oauthlib>=0.7
python-telegram-bot>=20.0
python-dotenv>=1.0
psycopg2-binary==2.9.11
bcrypt==5.0.0
python-telegram-bot-calendar==1.0.5
npm==0.1.1